+++ src/Service/S3/tests/Integration/S3ClientTest.php
@@ -36,6 +36,7 @@ use AsyncAws\S3\Input\PutBucketTaggingRequest;
 use AsyncAws\S3\Input\PutObjectAclRequest;
 use AsyncAws\S3\Input\PutObjectRequest;
 use AsyncAws\S3\Input\PutObjectTaggingRequest;
+use AsyncAws\S3\Input\PutPublicAccessBlockRequest;
 use AsyncAws\S3\Input\UploadPartCopyRequest;
 use AsyncAws\S3\Input\UploadPartRequest;
 use AsyncAws\S3\Result\PutObjectOutput;
@@ -55,6 +56,7 @@ use AsyncAws\S3\ValueObject\LambdaFunctionConfiguration;
 use AsyncAws\S3\ValueObject\NotificationConfiguration;
 use AsyncAws\S3\ValueObject\NotificationConfigurationFilter;
 use AsyncAws\S3\ValueObject\Owner;
+use AsyncAws\S3\ValueObject\PublicAccessBlockConfiguration;
 use AsyncAws\S3\ValueObject\QueueConfiguration;
 use AsyncAws\S3\ValueObject\S3KeyFilter;
 use AsyncAws\S3\ValueObject\Tag;
@@ -915,6 +917,29 @@ class S3ClientTest extends TestCase
         $result->resolve();
 
         self::assertNull($result->getVersionId());
+        self::assertSame(200, $result->info()['status']);
+    }
+
+    public function testPutPublicAccessBlock(): void
+    {
+        $client = $this->getClient();
+        $bucket = 'examplebucket';
+        $expectedBucketOwner = '0123456789';
+
+        $input = new PutPublicAccessBlockRequest([
+            'Bucket' => $bucket,
+            'PublicAccessBlockConfiguration' => new PublicAccessBlockConfiguration([
+                'BlockPublicAcls' => true,
+                'IgnorePublicAcls' => false,
+                'BlockPublicPolicy' => false,
+                'RestrictPublicBuckets' => true,
+            ]),
+            'ExpectedBucketOwner' => $expectedBucketOwner,
+        ]);
+        $result = $client->putPublicAccessBlock($input);
+
+        $result->resolve();
+
         self::assertSame(200, $result->info()['status']);
     }